+++ shiny/ui/fill/__init__.py
@@ -1,19 +1,15 @@
 from ._fill import (
-    as_fill_carrier,
     as_fill_item,
     as_fillable_container,
     remove_all_fill,
-    is_fill_carrier,
     is_fill_item,
     is_fillable_container,
 )
 
 __all__ = (
-    "as_fill_carrier",
     "as_fillable_container",
     "as_fill_item",
     "remove_all_fill",
-    "is_fill_carrier",
     "is_fillable_container",
     "is_fill_item",
 )